(PC-37115)[PRO] refactor: move SummaryScreen to pages/IndividualOffer/IndividualOfferSummary
