+++ pro/src/pages/IndividualOffer/IndividualOfferConfirmation/components/IndividualOfferConfirmationScreen.tsx
@@ -1,10 +1,10 @@
 import { GetIndividualOfferResponseModel } from 'apiClient/v1'
 import { OFFER_STATUS_PENDING } from 'commons/core/Offers/constants'
 import { formatDateTimeParts, isDateValid } from 'commons/utils/date'
-import { DisplayOfferInAppLink } from 'components/IndividualOffer/SummaryScreen/DisplayOfferInAppLink/DisplayOfferInAppLink'
 import fullLinkIcon from 'icons/full-link.svg'
 import fullValidateIcon from 'icons/full-validate.svg'
 import fullWaitIcon from 'icons/full-wait.svg'
+import { DisplayOfferInAppLink } from 'pages/IndividualOffer/IndividualOfferSummary/components/DisplayOfferInAppLink/DisplayOfferInAppLink'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
