+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/IndividualOfferSummary.tsx
@@ -5,8 +5,9 @@ import { OFFER_WIZARD_MODE } from 'commons/core/Offers/constants'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { IndividualOfferLayout } from 'components/IndividualOffer/IndividualOfferLayout/IndividualOfferLayout'
 import { getTitle } from 'components/IndividualOffer/IndividualOfferLayout/utils/getTitle'
-import { SummaryScreen } from 'components/IndividualOffer/SummaryScreen/SummaryScreen'
 import { Spinner } from 'ui-kit/Spinner/Spinner'
+
+import { IndividualOfferSummaryScreen } from './components/IndividualOfferSummaryScreen'
 
 const IndividualOfferSummary = (): JSX.Element | null => {
   const { offer, publishedOfferWithSameEAN } = useIndividualOfferContext()
@@ -33,7 +34,7 @@ const IndividualOfferSummary = (): JSX.Element | null => {
       mode={mode}
       venueHasPublishedOfferWithSameEan={Boolean(publishedOfferWithSameEAN)}
     >
-      <SummaryScreen />
+      <IndividualOfferSummaryScreen />
     </IndividualOfferLayout>
   )
 }
