+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/DisplayOfferInAppLink/DisplayOfferInAppLink.tsx
