+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/EventPublicationForm/__specs__/validationSchema.spec.ts
