+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/OfferSection/OfferSection.tsx
