+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/OfferSection/__specs__/serializer.spec.ts
