+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/OfferSection/serializer.ts
