+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/OfferSection/utils.ts
