+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/PriceCategoriesSection/PriceCategoriesSection.module.scss
