+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/PriceCategoriesSection/PriceCategoriesSection.tsx
