+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/PriceCategoriesSection/__specs__/PriceCategoriesSection.spec.tsx
