+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/RecurrenceSection/RecurrenceSection.tsx
