+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/RecurrenceSection/__specs__/RecurrenceSection.spec.tsx
