+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/StockSection.module.scss
