+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/StockSection.tsx
