+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/StockThingSection/StockThingSection.tsx
