+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/StockThingSection/__specs__/StockThingSection.spec.tsx
