+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/__specs__/StockSection.spec.tsx
