+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/StockSection/__specs__/getStockWarningText.spec.tsx
